Add behavior to "backspace" and "clear" buttons
When clicked, the "backspace" button pop out the
last number written on display.

When clicked, the "clear" button replaces
everything on display to value 0.

The number on display has been sized to be bigger
than before. This way the result can be seen more
easily.

Some variable names on script.js was changed to
add the "btn" prefix; it is easier to identify
each variable purpose this way.