Add functionality to operator buttons
Numbers are taken two by two and an operation
is done using them.

Everytime an operator button is clicked, the
value stored on display is caught as an operand
of the last expression mounted, if any, and the
result of this expression is used as an operand
for the next expression to be built using the
clicked operator.

Equality operator (=) has no functionality yet.
Its attribute in index.html was changed from class
"operator" to id "equality" so that it does not
behavior like other operators when clicked.

Since the remainder operator is present on
calculator's GUI, it's been added as a case
option on "operate" function.

Divide, multiplication and minus signs had their
value changed because they were not being
recognized by "operate" function in case statement.
It is easier to just throw their values into the
function than parse them before.